+++ src/instances/garbage/instance.ts
@@ -1,0 +1,7 @@
+import type { WorkbenchInstanceRegistration } from '../instance-types'
+import { garbageInstance } from './config'
+
+export const garbageWorkbenchInstance = {
+  instance: garbageInstance,
+  default: true,
+} satisfies WorkbenchInstanceRegistration