+++ src/instances/instance-types.ts
@@ -1,0 +1,7 @@
+import type { WorkbenchInstance, WorkbenchSnapshot } from '../core/workbench'
+
+export type WorkbenchInstanceRegistration = {
+  instance: WorkbenchInstance
+  default?: boolean
+  staticSnapshot?: () => WorkbenchSnapshot
+}